fix benchmark metric metadata and finalize audit cleanup
- correct benchmark metric helpers and bytes-per-iteration metadata

- align precision flags with the supported verification schema

- publish static signatures for timeout-prone lab benchmarks

- finalize scope classification and shared audit state updates

+++ code/ch13/optimized_bandwidth_naive.py
@@ -132,7 +132,7 @@ class OptimizedBandwidthCoalescedBenchmark(VerificationPayloadMixin, BaseBenchma
         return compute_precision_metrics(
             fp32_time_ms=None,
             reduced_precision_time_ms=getattr(self, '_last_elapsed_ms', None),
-            precision_type="fp8",
+            precision_type="fp32",
         )
 
     def validate_result(self) -> Optional[str]:
