+++ code/ch13/optimized_dataloader_default.py
@@ -183,7 +183,7 @@ class OptimizedDataloaderTunedBenchmark(VerificationPayloadMixin, BaseBenchmark)
         return compute_precision_metrics(
             fp32_time_ms=None,
             reduced_precision_time_ms=getattr(self, '_last_elapsed_ms', None),
-            precision_type="fp8",
+            precision_type="fp32",
         )
 
     def validate_result(self) -> Optional[str]:
