+++ code/ch19/baseline_vectorization_memory.py
@@ -92,7 +92,7 @@ class VectorizationBenchmark(VerificationPayloadMixin, BaseBenchmark):
         return compute_precision_metrics(
             fp32_time_ms=getattr(self, '_last_elapsed_ms', None),
             reduced_precision_time_ms=None,
-            precision_type="fp8",
+            precision_type="fp32",
         )
 
     def validate_result(self) -> Optional[str]:
